Removing rAzureBatch::deleteJob
Please, delete this pull request if I'm wrong and there is a real reason to delete the previous job. 
In my humble opinion, it is harder to debug R code execution on the node side after the jobs were deleted.
I still have not found the way to do this.